// apps/web/console/src/features/account/api/account.api.ts
export interface Address {
  id: number
  label: string | null
  type: string
  isDefault: boolean
  street1: string
  street2: string | null
  city: string
  state: string
  postalCode: string
  country: string
  phone: string | null
  createdAt: string
  updatedAt: string
}

export interface AddressInput {
  label?: string | null
  type?: 'shipping' | 'billing'
  street1: string
  street2?: string | null
  city: string
  state: string
  postalCode: string
  country: string
  phone?: string | null
  isDefault?: boolean
}